move override behave tests out of core
